src/requirements.txt
psutil
wmi
pywin32
customtkinter
packaging
pillow
yara-python
requests
pyyaml
